code cleanup, separated NewPackage vs UpdateExisting

+++ src/squirrel/squirrelDataDictionary.h
@@ -56,7 +56,7 @@ public:
     bool isValid() { return valid; }
     QString Error() { return err; }
     qint64 GetObjectID() { return objectID; }
-    void SetObjectID(int id) { objectID = id; }
+    void SetObjectID(qint64 id) { objectID = id; }
     QString VirtualPath();
 
     /* JSON elements */
