+++ src/squirrel/squirrelSubject.h
@@ -50,7 +50,7 @@ public:
     bool isValid() { return valid; }
     QString Error() { return err; }
     qint64 GetObjectID() { return objectID; }
-    void SetObjectID(int id) { objectID = id; }
+    void SetObjectID(qint64 id) { objectID = id; }
     void SetDirFormat(QString subject_DirFormat) {subjectDirFormat = subject_DirFormat; }
     QString VirtualPath();
